Improve performance of treewidth computation (esp. minor-min-width).
Here I used some information that we have to save on the number of
increase and decrease operations in the Fibonacci heap. These operations
currently require most time. Improves performance by roughly one order
of magnitude.

In the process I added a progress meter to get a sense of the speed of the
computation.